+++ R/grouped-dt.r
@@ -37,7 +37,6 @@ grouped_dt <- function(data, vars, copy = TRUE) {
   data
 }
 
-#' @export
 groups.grouped_dt <- function(x) {
   attr(x, "vars")
 }
@@ -47,7 +46,6 @@ groups.grouped_dt <- function(x) {
 #' @export
 is.grouped_dt <- function(x) inherits(x, "grouped_dt")
 
-#' @export
 print.grouped_dt <- function(x, ..., n = NULL, width = NULL) {
   cat("Source: local data table ", dplyr::dim_desc(x), "\n", sep = "")
   cat("Groups: ", commas(deparse_all(dplyr::groups(x))), "\n", sep = "")
@@ -56,27 +54,23 @@ print.grouped_dt <- function(x, ..., n = NULL, width = NULL) {
   invisible(x)
 }
 
-#' @importFrom dplyr group_size
-#' @export
 group_size.grouped_dt <- function(x) {
   dplyr::summarise_(x, n = ~n())$n
 }
 
-#' @importFrom dplyr n_groups
-#' @export
 n_groups.grouped_dt <- function(x) {
   nrow(dt_subset(x, , quote(list(1))))
 }
 
-#' @export
 #' @importFrom dplyr group_by_
+group_by.data.table <- function(.data, ..., add = FALSE) {
+  group_by_(.data, .dots = lazyeval::lazy_dots(...), add = add)
+}
 group_by_.data.table <- function(.data, ..., .dots, add = FALSE) {
   groups <- dplyr::group_by_prepare(.data, ..., .dots = .dots, add = add)
   grouped_dt(groups$data, groups$groups)
 }
 
-#' @export
-#' @importFrom dplyr ungroup
 ungroup.grouped_dt <- function(x, ...) {
   data.table::setattr(x, "vars", NULL)
   data.table::setattr(x, "class", setdiff(class(x), "grouped_dt"))
@@ -86,7 +80,6 @@ ungroup.grouped_dt <- function(x, ...) {
 
 # Do ---------------------------------------------------------------------------
 
-#' @export
 do_.grouped_dt <- function(.data, ..., .dots) {
   args <- lazyeval::all_dots(.dots, ...)
   env <- lazyeval::common_env(args)
@@ -130,7 +123,6 @@ named_args <- function(args) {
 
 # Set operations ---------------------------------------------------------------
 
-#' @export
 distinct_.grouped_dt <- function(.data, ..., .dots) {
   groups <- lazyeval::as.lazy_dots(groups(.data))
   dist <- distinct_vars(.data, ..., .dots = c(.dots, groups))