+++ R/joins.R
@@ -54,23 +54,15 @@ join_dt <- function(op) {
   # nocov end
 }
 
-#' @export
 #' @rdname join.tbl_dt
-#' @importFrom dplyr inner_join
 inner_join.data.table <- join_dt({merge(x, y, by = by$x, allow.cartesian = TRUE)})
 
-#' @export
-#' @importFrom dplyr left_join
 #' @rdname join.tbl_dt
 left_join.data.table  <- join_dt({merge(x, y, by = by$x, all.x = TRUE, allow.cartesian = TRUE)})
 
-#' @export
-#' @importFrom dplyr right_join
 #' @rdname join.tbl_dt
 right_join.data.table  <- join_dt(merge(x, y, by = by$x, all.y = TRUE, allow.cartesian = TRUE))
 
-#' @export
-#' @importFrom dplyr semi_join
 #' @rdname join.tbl_dt
 semi_join.data.table  <- join_dt({
   # http://stackoverflow.com/questions/18969420/perform-a-semi-join-with-data-table
@@ -79,13 +71,9 @@ semi_join.data.table  <- join_dt({
   x[w]
 })
 
-#' @export
-#' @importFrom dplyr anti_join
 #' @rdname join.tbl_dt
 anti_join.data.table <- join_dt({x[!y, allow.cartesian = TRUE]})
 
-#' @export
-#' @importFrom dplyr full_join
 #' @rdname join.tbl_dt
 # http://stackoverflow.com/a/15170956/946850
 full_join.data.table <- join_dt({merge(x, y, by = by$x, all = TRUE, allow.cartesian = TRUE)})