+++ R/sample.R
@@ -1,11 +1,8 @@
-#' @importFrom dplyr sample_n
-#' @export
 sample_n.tbl_dt <- function(tbl, size, replace = FALSE, weight = NULL,
                             .env = parent.frame()) {
   tbl_dt(NextMethod())
 }
 
-#' @export
 sample_n.grouped_dt <- function(tbl, size, replace = FALSE, weight = NULL,
                                 .env = parent.frame()) {
 
@@ -18,14 +15,11 @@ sample_n.grouped_dt <- function(tbl, size, replace = FALSE, weight = NULL,
   grouped_dt(tbl[idx], groups(tbl))
 }
 
-#' @importFrom dplyr sample_frac
-#' @export
 sample_frac.tbl_dt <- function(tbl, size = 1, replace = FALSE, weight = NULL,
                                .env = parent.frame()) {
   tbl_dt(NextMethod())
 }
 
-#' @export
 sample_frac.grouped_dt <- function(tbl, size = 1, replace = FALSE, weight = NULL,
                                    .env = parent.frame()) {