+++ R/sets.R
@@ -1,7 +1,10 @@
 
 # Set operations ---------------------------------------------------------------
 
-#' @export
+distinct.data.table <- function(.data, ..., .keep_all = FALSE) {
+  distinct_(.data, .dots = lazyeval::lazy_dots(...), .keep_all = .keep_all)
+}
+
 #' @importFrom dplyr distinct_
 distinct_.data.table <- function(.data, ..., .dots, .keep_all = FALSE) {
   dist <- distinct_vars(.data, ..., .dots = .dots)
@@ -18,8 +21,6 @@ distinct_.data.table <- function(.data, ..., .dots, .keep_all = FALSE) {
 
   res
 }
-
-#' @export
 distinct_.tbl_dt <- function(.data, ..., .dots) {
   tbl_dt(NextMethod(), copy = FALSE)
 }