+++ R/tbl-dt.r
@@ -152,18 +152,20 @@ and_expr <- function(exprs) {
   left
 }
 
+# The S3 method is registered manually in .onLoad() to avoid an R CMD
+# check warning
+
+filter.data.table <- function(.data, ...) {
+  filter_(.data, .dots = lazyeval::lazy_dots(...))
+}
+
 #' @importFrom dplyr filter_
-#' @export
 filter_.grouped_dt <- function(.data, ..., .dots) {
   grouped_dt(NextMethod(), groups(.data), copy = FALSE)
 }
-
-#' @export
 filter_.tbl_dt <- function(.data, ..., .dots) {
   tbl_dt(NextMethod(), copy = FALSE)
 }
-
-#' @export
 filter_.data.table <- function(.data, ..., .dots) {
   dots <- lazyeval::all_dots(.dots, ...)
   env <- lazyeval::common_env(dots)
@@ -178,18 +180,17 @@ filter_.data.table <- function(.data, ..., .dots) {
 
 # Summarise --------------------------------------------------------------------
 
+summarise.data.table <- function(.data, ...) {
+  summarise_(.data, .dots = lazyeval::lazy_dots(...))
+}
+
 #' @importFrom dplyr summarise_
-#' @export
 summarise_.grouped_dt <- function(.data, ..., .dots) {
   grouped_dt(NextMethod(), drop_last(groups(.data)), copy = FALSE)
 }
-
-#' @export
 summarise_.tbl_dt <- function(.data, ..., .dots) {
   tbl_dt(NextMethod(), copy = FALSE)
 }
-
-#' @export
 summarise_.data.table <- function(.data, ..., .dots) {
   dots <- lazyeval::all_dots(.dots, ..., all_named = TRUE)
 
@@ -199,18 +200,17 @@ summarise_.data.table <- function(.data, ..., .dots) {
 
 # Mutate -----------------------------------------------------------------------
 
+mutate.data.table <- function(.data, ...) {
+  mutate_(.data, .dots = lazyeval::lazy_dots(...))
+}
+
 #' @importFrom dplyr mutate_
-#' @export
 mutate_.grouped_dt <- function(.data, ..., .dots) {
   grouped_dt(NextMethod(), drop_last(groups(.data)), copy = FALSE)
 }
-
-#' @export
 mutate_.tbl_dt <- function(.data, ..., .dots) {
   tbl_dt(NextMethod(), copy = FALSE)
 }
-
-#' @export
 mutate_.data.table <- function(.data, ..., .dots) {
   dots <- lazyeval::all_dots(.dots, ..., all_named = TRUE)
   names <- lapply(names(dots), as.name)
@@ -230,18 +230,17 @@ mutate_.data.table <- function(.data, ..., .dots) {
 
 # Arrange ----------------------------------------------------------------------
 
+arrange.data.table <- function(.data, ...) {
+  arrange_(.data, .dots = lazyeval::lazy_dots(...))
+}
+
 #' @importFrom dplyr arrange_
-#' @export
 arrange_.grouped_dt <- function(.data, ..., .dots) {
   grouped_dt(NextMethod(), groups(.data), copy = FALSE)
 }
-
-#' @export
 arrange_.tbl_dt <- function(.data, ..., .dots) {
   tbl_dt(NextMethod(), copy = FALSE)
 }
-
-#' @export
 arrange_.data.table <- function(.data, ..., .dots) {
   dots <- lazyeval::all_dots(.dots, ...)
 
@@ -254,8 +253,11 @@ arrange_.data.table <- function(.data, ..., .dots) {
 
 # Select -----------------------------------------------------------------------
 
+select.data.table <- function(.data, ...) {
+  select_(.data, .dots = lazyeval::lazy_dots(...))
+}
+
 #' @importFrom dplyr select_
-#' @export
 select_.grouped_dt <- function(.data, ..., .dots) {
   dots <- lazyeval::all_dots(.dots, ...)
   vars <- dplyr::select_vars_(names(.data), dots,
@@ -265,8 +267,6 @@ select_.grouped_dt <- function(.data, ..., .dots) {
 
   grouped_dt(out, groups(.data), copy = FALSE)
 }
-
-#' @export
 select_.data.table <- function(.data, ..., .dots) {
   dots <- lazyeval::all_dots(.dots, ...)
   vars <- dplyr::select_vars_(names(.data), dots)
@@ -275,16 +275,17 @@ select_.data.table <- function(.data, ..., .dots) {
   data.table::setnames(out, names(vars))
   out
 }
-
-#' @export
 select_.tbl_dt <- function(.data, ..., .dots) {
   tbl_dt(NextMethod(), copy = FALSE)
 }
 
 # Rename -----------------------------------------------------------------------
 
+rename.data.table <- function(.data, ...) {
+  rename_(.data, .dots = lazyeval::lazy_dots(...))
+}
+
 #' @importFrom dplyr rename_
-#' @export
 rename_.grouped_dt <- function(.data, ..., .dots) {
   dots <- lazyeval::all_dots(.dots, ...)
   vars <- dplyr::rename_vars_(names(.data), dots)
@@ -294,8 +295,6 @@ rename_.grouped_dt <- function(.data, ..., .dots) {
 
   grouped_dt(out, groups(.data), copy = FALSE)
 }
-
-#' @export
 rename_.data.table <- function(.data, ..., .dots) {
   dots <- lazyeval::all_dots(.dots, ...)
   vars <- dplyr::rename_vars_(names(.data), dots)
@@ -304,8 +303,6 @@ rename_.data.table <- function(.data, ..., .dots) {
   data.table::setnames(out, names(vars))
   out
 }
-
-#' @export
 rename_.tbl_dt <- function(.data, ..., .dots) {
   tbl_dt(NextMethod(), copy = FALSE)
 }
@@ -313,18 +310,17 @@ rename_.tbl_dt <- function(.data, ..., .dots) {
 
 # Slice -------------------------------------------------------------------
 
+slice.data.table <- function(.data, ...) {
+  slice_(.data, .dots = lazyeval::lazy_dots(...))
+}
+
 #' @importFrom dplyr slice_
-#' @export
 slice_.grouped_dt <- function(.data, ..., .dots) {
   grouped_dt(NextMethod(), groups(.data), copy = FALSE)
 }
-
-#' @export
 slice_.tbl_dt <- function(.data, ..., .dots) {
   tbl_dt(NextMethod(), copy = FALSE)
 }
-
-#' @export
 slice_.data.table <- function(.data, ..., .dots) {
   dots <- lazyeval::all_dots(.dots, ..., all_named = TRUE)
   env <- lazyeval::common_env(dots)