+++ tests/testthat/helper-data.R
@@ -1,0 +1,4 @@
+
+lahman_dt <- function() {
+  dtplyr::src_dt("Lahman")
+}