+++ NEWS.md
@@ -1,4 +1,8 @@
-# dtplyr 0.0.1.9000
+# dtplyr 0.0.2
+
+- This is a compatibility release. It makes dtplyr compatible with
+  dplyr 0.6.0 in addition to dplyr 0.5.0.
+
 
 # dtplyr 0.0.1